Gate later research pods behind progression and price APIs against similar-capability rivals.
Starter Foundations-only pods, office HQ kits, and radar/plans comparables keep later methods locked and demand scored against peers instead of every public model.

+++ src/sim/systems/hosting.ts
@@ -172,7 +172,7 @@ export function modelHostNeed(
   m: Model,
   opts?: { precision?: ServePrecision; concurrentRequests?: number; contextTokens?: number },
 ): ModelHostNeed {
-  const precision = opts?.precision ?? "fp16";
+  const precision = opts?.precision ?? "fp32";
   const memory = estimateServingMemory({
     model: m,
     precision,
